Move tv series methods to TvSeriesDataService
